Move templates into own file
